+++ src/main/java/com/github/wintersteve25/tau/components/Text.java
@@ -1,8 +1,13 @@
 package com.github.wintersteve25.tau.components;
 
+import com.github.wintersteve25.tau.build.UIBuilder;
+import com.github.wintersteve25.tau.utils.RenderProvider;
+import com.github.wintersteve25.tau.utils.Size;
 import com.mojang.blaze3d.matrix.MatrixStack;
 
+import com.sun.jna.platform.win32.Wdm;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.AbstractGui;
 import net.minecraft.client.gui.FontRenderer;
 import net.minecraft.client.gui.IGuiEventListener;
 import net.minecraft.client.gui.IRenderable;
@@ -18,7 +23,9 @@ import com.github.wintersteve25.tau.utils.Vector2i;
 
 import java.util.List;
 
-public final class Text implements PrimitiveUIComponent {
+public final class Text implements PrimitiveUIComponent, RenderProvider {
+    
+    private static int ellipsisWidth = 0;
     
     private final ITextComponent text;
     private final Color color;
@@ -32,30 +39,52 @@ public final class Text implements PrimitiveUIComponent {
 
     @Override
     public Vector2i build(Layout layout, List<IRenderable> renderables, List<DynamicUIComponent> dynamicUIComponents, List<IGuiEventListener> eventListeners) {
-        int width = Minecraft.getInstance().font.width(text);
-        int height = overflowBehaviour == OverflowBehaviour.WRAP ?
-                Minecraft.getInstance().font.wordWrapHeight(text.getContents(), layout.getMaximumLength(Axis.HORIZONTAL)) :
+        FontRenderer fontRenderer = Minecraft.getInstance().font;
+        int width = fontRenderer.width(text);
+        ellipsisWidth = fontRenderer.width("...");
+        
+        boolean willOverflow = width > layout.getWidth();
+        if (willOverflow) {
+            width = layout.getWidth();
+        }
+        
+        int height = willOverflow && overflowBehaviour == OverflowBehaviour.WRAP ?
+                fontRenderer.wordWrapHeight(text.getContents(), width) :
                 9; // constant for line height in minecraft
+        
         int x = layout.getPosition(Axis.HORIZONTAL, width);
         int y = layout.getPosition(Axis.VERTICAL, height);
 
-        renderables.add((pMatrixStack, pMouseX, pMouseY, pPartialTicks) -> render(pMatrixStack, x, y, width));
-    
+        int finalWidth = width;
+        if (overflowBehaviour != OverflowBehaviour.CLIP) {
+            renderables.add((pMatrixStack, pMouseX, pMouseY, pPartialTicks) -> render(pMatrixStack, pMouseX, pMouseY, pPartialTicks, x, y, finalWidth, height));
+        } else {
+            UIBuilder.build(
+                new Layout(width, height, x, y),
+                new Clip.Builder()
+                    .build(new Render(this)), 
+                renderables,
+                dynamicUIComponents, 
+                eventListeners
+            );
+        }
+        
         return new Vector2i(width, height);
     }
-
-    private void render(MatrixStack matrixStack, int xPosition, int yPosition, int width) {
+    
+    @Override
+    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks, int x, int y, int width, int height) {
         FontRenderer fontRenderer = Minecraft.getInstance().font;
 
         switch (overflowBehaviour) {
-            case OVERFLOW:
-                fontRenderer.drawShadow(matrixStack, text.getString(), xPosition, yPosition, color.getAARRGGBB(), color.hasTransparency());
-                break;
             case WRAP:
-                fontRenderer.drawWordWrap(text, xPosition, yPosition, width, color.getAARRGGBB());
+                fontRenderer.drawWordWrap(text, x, y, width, color.getAARRGGBB());
                 break;
-            case CLIP:
-                fontRenderer.drawShadow(matrixStack, fontRenderer.substrByWidth(text, width).getString(), xPosition, yPosition, color.getAARRGGBB(), color.hasTransparency());
+            case ELLIPSIS:
+                fontRenderer.drawShadow(matrixStack, fontRenderer.substrByWidth(text, width - ellipsisWidth).getString() + "...", x, y, color.getAARRGGBB(), color.hasTransparency());
+                break;
+            default:
+                fontRenderer.drawShadow(matrixStack, text.getString(), x, y, color.getAARRGGBB(), color.hasTransparency());
                 break;
         }
     }
@@ -101,5 +130,6 @@ public final class Text implements PrimitiveUIComponent {
         OVERFLOW,
         WRAP,
         CLIP,
+        ELLIPSIS
     }
 }