+++ src/main/java/com/github/wintersteve25/tau/tests/TestSized.java
@@ -1,11 +1,19 @@
 package com.github.wintersteve25.tau.tests;
 
+import com.github.wintersteve25.tau.components.Button;
+import com.github.wintersteve25.tau.components.Center;
+import com.github.wintersteve25.tau.components.Sized;
+import com.github.wintersteve25.tau.components.Text;
 import com.github.wintersteve25.tau.components.base.UIComponent;
 import com.github.wintersteve25.tau.layout.Layout;
+import com.github.wintersteve25.tau.utils.Size;
 
 public class TestSized implements UIComponent {
     @Override
     public UIComponent build(Layout layout) {
-        return null;
+        return new Sized(
+                Size.percentage(0.2f, 0.05f),
+                new Button.Builder().build(new Center(new Text.Builder("Hello")))
+        );
     }
 }