+++ src/main/java/com/github/wintersteve25/tau/tests/TestText.java
@@ -1,11 +1,18 @@
 package com.github.wintersteve25.tau.tests;
 
+import com.github.wintersteve25.tau.components.Center;
+import com.github.wintersteve25.tau.components.Sized;
+import com.github.wintersteve25.tau.components.Text;
 import com.github.wintersteve25.tau.components.base.UIComponent;
 import com.github.wintersteve25.tau.layout.Layout;
+import com.github.wintersteve25.tau.utils.Size;
 
 public class TestText implements UIComponent {
     @Override
     public UIComponent build(Layout layout) {
-        return null;
+        return new Center(
+            new Text.Builder("Wow hello a lot of textttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttt")
+                    .withOverflowBehaviour(Text.OverflowBehaviour.ELLIPSIS)
+        );
     }
 }