+++ src/main/java/com/github/wintersteve25/tau/utils/Color.java
@@ -12,11 +12,16 @@ public final class Color {
 
     public Color(int hex) {
         this.hex = hex;
-        this.hasTransparency = false;
+        this.hasTransparency = ((hex >> 24) & 0xFF) < 255;
+    }
+    
+    private Color(int hex, boolean hasTransparency) {
+        this.hex = hex;
+        this.hasTransparency = hasTransparency;
     }
     
     public static Color fromRGBA(int r, int g, int b, int a) {
-        return new Color((a << 24) + (r << 16) + (g << 8) + b);
+        return new Color((a << 24) + (r << 16) + (g << 8) + b, a < 255);
     }
 
     public boolean hasTransparency() {